Circle file merged into distributions, seed can be set, Dokumentation improved, Exception raised if agents could not be placed
